+++ wrapper/main.js
@@ -1,39 +1,32 @@
-// Core part of Wrapper: Offline
-const RPC = require("discord-rpc");
+/***
+ * Wrapper: Offline
+ */
+// start server
 require("./server");
 
-
-// Loads env.json for Wrapper version and build number
-const env = Object.assign(process.env,
-	require('./env'));
-// env.json variables
-let version = env.WRAPPER_VER;
-let build = env.WRAPPER_BLD;
-
-
-// Discord rich presence
+/**
+ * rich presence
+ */
+// get version number
+const version = process.env.WRAPPER_VER;
 const rpc = new RPC.Client({
 	transport: "ipc"
 });
+
+// sets rpc activity when ready
 rpc.on("ready", () => {
-	// Sets RPC activity
 	rpc.setActivity({
-		// state: "Video List",
-		// disabled until automatic rpc status is done
-		details: "Version " + version +", build " + build,
+		details: `Version ${version}`,
 		startTimestamp: new Date(),
 		largeImageKey: "icon",
 		largeImageText: "Wrapper: Offline",
 		smallImageKey: "Wrapper: Offline",
 		smallImagetext: "Wrapper: Offline",
 	});
-	// Logs "Rich presence is on!" in the console
-	console.log("Rich presence is on!")
 });
-// Connects RPC to app
-rpc.login({
-	clientId: "866340172874383370"
-}).catch(
-	console.log('RPC connection failed.')
-);
+
+// connect rpc to app
+rpc
+	.login({ clientId: "866340172874383370" })
+	.catch(console.log("RPC connection failed."));