+++ changelog.md
@@ -2,6 +2,13 @@
 # Wrapper: Offline Changelog
 This changelog is based off [Keep a Changelog](https://keepachangelog.com) and ~~somewhat but not really~~ adheres to [Semantic Versioning](https://semver.org/). If you have no idea what that means, basically this changelog is formatted in a specific consistent way, and version numbers are formatted as X.Y.Z, X being a major not-backwards-compatible update, Y being a feature update, and X being bug fixes.
 
+## Version 1.3.0 - 2022-00-00 - a whole year wasted
+### Added
+### Changed
+  - Rich presence now uses process.env instead of reassigning env.json to process.env
+  - Removed build numbering
+### Fixed
+
 ## Version 1.2.3 - 2021-03-06 - i am HUNGRY FOR UPDATES!!!
 ### Added
   - Character browser